+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Shiva API:
   * Added endpoints `/api/random/artist`, `/api/random/album`, `/api/random/track`
     [#51](https://github.com/owncloud/music/issues/51)
+  * Added endpoint `/api/tracks/{id}/lyrics`
+    [#48](https://github.com/owncloud/music/issues/48)
   * Added endpoint `/api/whatsnew`
   * Systematically use plurals in endpoint names to match the specification: `api/artists/{id}` instead of `api/artist/{id}` etc.
 - Proprietary REST API: